% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_create_slides.R
\name{add_bullet_plus_chart_slide}
\alias{add_bullet_plus_chart_slide}
\title{Create Slide with Full Bullets on left and chart on the left}
\usage{
add_bullet_plus_chart_slide(
  p,
  p.title = " ",
  p.caption = " ",
  p.bullet = " ",
  p.chart
)
}
\arguments{
\item{p}{officer presentation object}

\item{p.title}{Title for slide}

\item{p.caption}{Text to be used for summary above the chart}

\item{p.bullet}{Text for bullet points}

\item{p.chart}{Chart object to display in slide}
}
\value{
Full Data slide with caption for PowerPoint presentation
}
\description{
This function creates a data slide with a bullets and a chart.
}
