% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_create_slides.R
\name{add_cover_slide}
\alias{add_cover_slide}
\title{Create Cover Slide}
\usage{
add_cover_slide(
  p = officer::read_pptx(system.file("extdata", "psrc-trends-template.pptx", package =
    "psrcslides")),
  p.img = system.file("extdata", "bellevuelakewashington.jpg", package = "psrcslides"),
  p.title = "Presentation Title",
  p.mtg = "Meeting Name",
  p.date = "Meeting Date"
)
}
\arguments{
\item{p}{PowerPoint Template File to use to create presentation - defaults to template installed with package}

\item{p.img}{Location or URL for background image for cover slide - ex "images/bellevuelakewashington_3.jpg"}

\item{p.title}{Title for the presentation}

\item{p.mtg}{Meeting or presentation venue}

\item{p.date}{Meeting date}
}
\value{
Cover slide for PowerPoint Presentation
}
\description{
This function allows you to create a title slide for a new presentation.
}
\examples{

psrc_pres <- add_cover_slide(p.title="Puget Sound Trends",
                             p.mtg="Growth Management Policy Board",
                             p.date="September 2022")

}
