% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_create_slides.R
\name{add_full_chart_slide}
\alias{add_full_chart_slide}
\title{Create Slide with Full width picture or image}
\usage{
add_full_chart_slide(p, p.title = "Slide Title", p.chart)
}
\arguments{
\item{p}{officer presentation object}

\item{p.title}{Title for slide}

\item{p.chart}{Chart object to display in slide}
}
\value{
Full Data slide with caption for PowerPoint presentation
}
\description{
This function creates a data slide with a full width object without caption text.
}
