% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrc_create_slides.R
\name{add_transition_slide}
\alias{add_transition_slide}
\title{Create Transition Slide}
\usage{
add_transition_slide(p, p.img, p.title = "Slide Title")
}
\arguments{
\item{p}{officer presentation file to add slide}

\item{p.img}{Location or URL for background image - ex "images/bellevuelakewashington_3.jpg"}

\item{p.title}{Title for slide}
}
\value{
Transition slide for PowerPoint Presentation
}
\description{
This function creates a transition slide with a background picture and white centered text.
}
\examples{

library(officer)

psrc_pres = read_pptx(system.file('extdata', 
                                  'psrc-trends-template.pptx', 
                                   package='psrcslides'))

psrc_pres <- add_transition_slide(p=psrc_pres, 
                                  p.title="Population Trends",
                                  p.img=system.file('extdata', 
                                                    'uw-equity.png',
                                                     package='psrcslides'))
                                  
}
