% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrcslides_details.R
\docType{package}
\name{psrcslides}
\alias{psrcslides}
\alias{psrcslides-package}
\title{psrcslides: A package for easy creation of PowerPoint Slide for the Puget Sound Region.}
\description{
Functions used to create PSRC PowerPoint slides using officer and a PSRC slide template.
This package is currently under development and will be updated frequently over the last half of 2022.
Colors and fonts are all consistent with design standards used by PSRC.
}
\author{
Craig Helmann \email{chelmann@psrc.org}
}
